refactoring code organization
